compress flow from test prog

+++ switch_to_fill/inc/router_request_args.h
@@ -17,7 +17,7 @@ typedef struct _timed_offload_request_args{
   ax_comp *comp;
   struct hw_desc *desc;
   char *src_payload;
-  int src_size;
+  uint64_t src_size;
   char *dst_payload;
   int id;
   uint64_t *ts0;
